+++ src/Groups/GroupDetails/utils.js
@@ -1,15 +1,8 @@
 import { LIMIT_MAX } from '@folio/stripes-acq-components';
 
-const emptyGroupSummary = {
-  allocated: 0,
-  unavailable: 0,
-  available: 0,
-};
-
-// eslint-disable-next-line import/prefer-default-export
 export const getGroupSummary = (groupSummariesMutator, groupId, fiscalYearId) => {
   if (!fiscalYearId) {
-    return Promise.resolve(emptyGroupSummary);
+    return Promise.resolve({});
   }
 
   return groupSummariesMutator.GET({
@@ -19,6 +12,6 @@ export const getGroupSummary = (groupSummariesMutator, groupId, fiscalYearId) =>
     },
   })
     .then(groupSummaries => {
-      return groupSummaries[0] || emptyGroupSummary;
+      return groupSummaries[0] || {};
     });
 };